# contract/openapi.yaml
openapi: 3.0.3
info:
  license:
    name: MIT
  title: goERP API
  version: 1.0.0
  description: |
    Multi-tenant ERP REST API (/api/v1).

    ## Conventions (as implemented — this spec documents reality, not intent)
    - Success envelope: most responses are `{"data": <payload>}`.
      **Exceptions:** single-item customer and supplier operations return the object bare,
      without the `data` wrapper.
    - Paginated lists: `{"data": [...], "pagination": {"limit","offset","total"}}`.
    - Errors: `{"error": {"code": "...", "message": "...", "details": {...}?}}`.
    - Money is a decimal **string** with 2 fraction digits (`"1234.56"`), never a JSON number.
    - Timestamps are RFC3339 strings; report dates are `YYYY-MM-DD`.
    - All business routes require `Authorization: Bearer <accessToken>` and are scoped to the
      tenant embedded in the JWT. Cross-tenant rows are invisible.
    - Role matrix: `admin > manager > operator > viewer`.
    - Deletes are soft (`deleted_at` set); soft-deleted rows read as 404 afterwards.

    ## Global error codes
    | Code | Status | Meaning |
    |------|--------|---------|
    | UNAUTHORIZED | 401 | missing/invalid bearer token, missing tenant context, bad credentials, missing/invalid refresh token |
    | FORBIDDEN | 403 | authenticated but role insufficient for the route |
    | RATE_LIMITED | 429 | login rate limit exceeded (5 attempts / 15 min per source IP) |
    | INVALID_JSON | 400 | request body is not valid JSON |
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /auth/refresh:
    post:
      tags: [Auth]
      security: []
      operationId: refreshToken
      summary: Rotate refresh token and issue a new access token
      description: |
        Reads the `refresh_token` HttpOnly cookie (request body ignored). Refresh tokens are
        single-use: the old session is revoked and a new cookie is set. On failure the cookie
        is cleared.
      responses:
        '200':
          description: New access token; rotated refresh cookie
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LoginResponse' }
        '401':
          description: Cookie missing or session invalid
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
  /reports/sales-by-day:
    get:
      tags: [Reports]
      security: [{ bearerAuth: [] }]
      operationId: salesByDay
      summary: Confirmed-sale counts and totals per day
      x-roles: admin,manager
      parameters:
        - $ref: '#/components/parameters/FromQ'
        - $ref: '#/components/parameters/ToQ'
      responses:
        '200':
          description: One row per day (possibly empty)
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/SalesByDayPoint' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /reports/top-products:
    get:
      tags: [Reports]
      security: [{ bearerAuth: [] }]
      operationId: topProducts
      summary: Best-selling products in range
      x-roles: admin,manager
      parameters:
        - $ref: '#/components/parameters/FromQ'
        - $ref: '#/components/parameters/ToQ'
        - name: limit
          in: query
          required: false
          schema: { type: integer, default: 5 }
          description: <=0 or absent falls back to 5; no upper cap.
      responses:
        '200':
          description: Ranked rows (possibly empty)
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/TopProductRow' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /reports/sales-summary.pdf:
    get:
      tags: [Reports]
      security: [{ bearerAuth: [] }]
      operationId: salesSummaryPdf
      summary: PDF summary of confirmed sales in range
      x-roles: admin,manager
      parameters:
        - $ref: '#/components/parameters/FromQ'
        - $ref: '#/components/parameters/ToQ'
      responses:
        '200':
          description: Binary PDF (attachment vendas-resumo.pdf)
          content:
            application/pdf:
              schema: { type: string, format: binary }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /reports/sales/{id}/pdf:
    get:
      tags: [Reports]
      security: [{ bearerAuth: [] }]
      operationId: salePdf
      summary: PDF order document for one sale
      description: Broadest report role set — viewers may download single-sale PDFs.
      x-roles: admin,manager,operator,viewer
      parameters:
        - $ref: '#/components/parameters/SaleIdParam'
      responses:
        '200':
          description: Binary PDF (attachment venda-{id}.pdf)
          content:
            application/pdf:
              schema: { type: string, format: binary }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    SaleIdParam:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
      description: Sale UUID; unparseable behaves as a miss (404).
...
    FromQ:
      name: from
      in: query
      required: false
      schema: { type: string }
      description: Range start; RFC3339 datetime or YYYY-MM-DD. Defaults to now(UTC) minus 30 days; unparsable silently falls back.
    ToQ:
      name: to
      in: query
      required: false
      schema: { type: string }
      description: Range end; RFC3339 datetime or YYYY-MM-DD. Defaults to now(UTC); unparsable silently falls back.
  responses:
    Unauthorized:
      description: Missing/invalid bearer token or tenant context (UNAUTHORIZED)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Forbidden:
      description: Role not permitted on this route (FORBIDDEN)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    NotFound:
      description: Entity missing, soft-deleted, or outside the tenant (NOT_FOUND)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
    InternalError:
      description: Unexpected failure (INTERNAL_ERROR)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  schemas:
    ErrorBody:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
        details:
          type: object
          additionalProperties: { type: string }
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error: { $ref: '#/components/schemas/ErrorBody' }
...
    AuthUser:
      type: object
      required: [id, tenantId, email, name, role]
      properties:
        id: { type: string, format: uuid }
        tenantId: { type: string, format: uuid }
        email: { type: string, format: email }
        name: { type: string }
        role: { $ref: '#/components/schemas/Role' }
    LoginResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [accessToken, user]
          properties:
            accessToken: { type: string }
            user: { $ref: '#/components/schemas/AuthUser' }
    Role:
      type: string
      enum: [admin, manager, operator, viewer]
...
    SalesByDayPoint:
      type: object
      required: [date, count, total]
      properties:
        date: { type: string, format: date }
        count: { type: integer, format: int64 }
        total: { type: string, description: Decimal money as string. }
    TopProductRow:
      type: object
      required: [productName, productSku, quantity, revenue]
      properties:
        productName: { type: string }
        productSku: { type: string }
        quantity: { type: integer, format: int64 }
        revenue: { type: string, description: Decimal money as string. }
...